+++ rollup.config.js
@@ -6,7 +6,6 @@ import cleanup from 'rollup-plugin-cleanup';
 import generateHtml from 'rollup-plugin-generate-html';
 import alias from '@rollup/plugin-alias';
 import { dts } from 'rollup-plugin-dts';
-import path from 'path';
 import terser from '@rollup/plugin-terser';
 
 const BUILD_DIST = 'dist';
@@ -19,6 +18,7 @@ const commonPlugins = [
     babel({
         extensions: ['.js', '.ts'],
         babelHelpers: 'runtime',
+        plugins: ['@babel/plugin-transform-runtime'],
     }),
 ];
 
@@ -108,23 +108,15 @@ const entryPoints = {
     'converters/index': 'src/converters/index.ts',
 };
 
-const scriptletsCjsAndEsmConfig = {
+const scriptletsConfig = {
     input: entryPoints,
     output: [
         {
-            dir: `${BUILD_DIST}/cjs`,
-            format: 'cjs',
-            entryFileNames: '[name].js',
-            exports: 'named',
-            preserveModules: true,
-            preserveModulesRoot: 'src',
-        },
-        {
-            dir: `${BUILD_DIST}/esm`,
+            dir: BUILD_DIST,
             format: 'esm',
-            entryFileNames: '[name].mjs',
+            entryFileNames: '[name].js',
+            chunkFileNames: 'common/[name].js',
             exports: 'named',
-            preserveModules: true,
             preserveModulesRoot: 'src',
         },
     ],
@@ -137,12 +129,16 @@ const scriptletsCjsAndEsmConfig = {
          * ```
          */
         propertyReadSideEffects: false,
+
+        /**
+         * Assume that all modules do not have side effects
+         */
+        moduleSideEffects: false,
     },
     external: (id) => {
         return (
             // Added because when agtree is linked using 'yarn link', its ID does not contain 'node_modules'
-            id.startsWith('@babel/runtime')
-            || id.startsWith('js-yaml')
+            id.startsWith('js-yaml')
             || id.startsWith('@adguard/agtree')
         );
     },
@@ -162,7 +158,7 @@ const scriptletsCjsAndEsmConfig = {
             entries: [
                 {
                     find: 'scriptlets-func',
-                    replacement: path.resolve(__dirname, 'tmp/scriptlets-func.js'),
+                    replacement: 'tmp/scriptlets-func.js',
                 },
             ],
         }),
@@ -184,17 +180,17 @@ const typesConfig = {
             entries: [
                 {
                     find: 'scriptlets-func',
-                    replacement: path.resolve(__dirname, 'tmp/scriptlets-func.d.ts'),
+                    replacement: 'tmp/scriptlets-func.d.ts',
                 },
             ],
         }),
     ],
 };
 
-const scriptletsCjsAndEsm = [scriptletsCjsAndEsmConfig, typesConfig];
+const scriptlets = [scriptletsConfig, typesConfig];
 
 export {
-    scriptletsCjsAndEsm,
+    scriptlets,
     scriptletsListConfig,
     redirectsListConfig,
     click2LoadConfig,